Warnings clean-up, unnecessary casts.
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@542335 13f79535-47bb-0310-9956-ffa450edef68

+++ core-unit/src/test/java/org/apache/directory/server/core/prefs/ServerSystemPreferencesITest.java
@@ -127,7 +127,7 @@ public class ServerSystemPreferencesITest extends AbstractAdminTestCase
         testNode.putFloat( "cn", ( float ) 9.20 );
         testNode.sync();
         testNode = prefs.node( "testNode" );
-        assertTrue( ( float ) 9.20 == ( float ) testNode.getFloat( "cn", ( float ) 9.20 ) );
+        assertTrue( ( float ) 9.20 == testNode.getFloat( "cn", ( float ) 9.20 ) );
     }
